fix methods of compound controller/service/dao

+++ src/main/java/ua/netcracker/group3/automaticallytesting/dao/CompoundDAO.java
@@ -21,9 +21,9 @@ public interface CompoundDAO {
 
     List<CompoundActionWithActionIdAndPriority> findAllCompoundActionsWithActionIdAndPriorityByCompoundId(long compoundId);
 
-    List<ActionDtoWithIdNameVoid> findAllCompoundActionsByCompoundId(int compoundId);
+    List<ActionDtoWithIdNameVoid> findAllCompoundActionsByCompoundId(long compoundId);
 
-    void archiveCompoundById(int id);
+    void archiveCompoundById(long id);
 
     boolean checkIfNameExist(String name);
